requirements.txt
asgiref==3.8.1
Django==5.1.3
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
